move /user routes into userRoutes
